+++ application/backend/src/main/java/pl/ds/kyanite/components/helpers/PageModel.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package pl.ds.kyanite.components.models;
+package pl.ds.kyanite.components.helpers;
 
 import java.util.ArrayList;
 import java.util.List;
