+++ application/backend/src/main/java/pl/ds/kyanite/components/models/PageComponent.java
@@ -20,7 +20,7 @@ import org.apache.sling.api.resource.Resource;
 import org.apache.sling.models.annotations.DefaultInjectionStrategy;
 import org.apache.sling.models.annotations.Model;
 import org.apache.sling.models.annotations.injectorspecific.SlingObject;
-import pl.ds.kyanite.utils.PageUtil;
+import pl.ds.kyanite.components.utils.PageUtil;
 
 @Model(adaptables = Resource.class,
     defaultInjectionStrategy = DefaultInjectionStrategy.OPTIONAL)
