+++ application/backend/src/main/java/pl/ds/kyanite/components/utils/StringUtils.java
@@ -14,7 +14,22 @@
  * limitations under the License.
  */
 
-@Version("1.0.0")
-package pl.ds.kyanite.utils;
+package pl.ds.kyanite.components.utils;
 
-import org.osgi.annotation.versioning.Version;
+import java.util.Arrays;
+import java.util.List;
+
+public class StringUtils {
+
+  public static List<String> encryptEmail(String email) {
+    int atIndex = email.indexOf("@");
+    int dotIndex = email.lastIndexOf(".");
+    if (atIndex < 0 || dotIndex < 0) {
+      return List.of();
+    }
+    String emailPart1 = email.substring(0, atIndex);
+    String emailPart2 = email.substring(atIndex + 1, dotIndex);
+    String emailPart3 = email.substring(dotIndex + 1);
+    return Arrays.asList(emailPart1, emailPart2, emailPart3);
+  }
+}
