--- application/backend/src/main/java/pl/ds/kyanite/utils/StringUtils.java
@@ -1,35 +1,0 @@
-/*
- * Copyright (C) 2023 Dynamic Solutions
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package pl.ds.kyanite.utils;
-
-import java.util.Arrays;
-import java.util.List;
-
-public class StringUtils {
-
-  public static List<String> encryptEmail(String email) {
-    int atIndex = email.indexOf("@");
-    int dotIndex = email.lastIndexOf(".");
-    if (atIndex < 0 || dotIndex < 0) {
-      return List.of();
-    }
-    String emailPart1 = email.substring(0, atIndex);
-    String emailPart2 = email.substring(atIndex + 1, dotIndex);
-    String emailPart3 = email.substring(dotIndex + 1);
-    return Arrays.asList(emailPart1, emailPart2, emailPart3);
-  }
-}
