+++ application/backend/src/main/resources/apps/wcm/dialogs/components/select/select.json.html
@@ -14,7 +14,7 @@
     limitations under the License.
 */-->
 
-<sly data-sly-use.model="pl.ds.kyanite.components.models.SelectModel"></sly>
+<sly data-sly-use.model="pl.ds.kyanite.components.helpers.SelectModel"></sly>
 {
     "type": "/apps/websight-dialogs-view/web-resources/components/Select.js",
     "props": {
