+++ application/backend/src/main/resources/libs/kyanite/components/buttons/.content.json
@@ -1,7 +1,10 @@
 {
   "isContainer": true,
   "group": "Kyanite Elements",
-  "allowedComponents": ["/libs/kyanite/components/button"],
+  "allowedComponents": [
+    "/apps/kyanite/components/button",
+    "/libs/kyanite/components/button"
+  ],
   "sling:resourceType": "ws:Component",
   "title": "Buttons"
 }
