+++ application/backend/src/main/resources/libs/kyanite/components/page/body-content.html
@@ -25,9 +25,9 @@
 
 <script src="/libs/kyanite/web_root/main.js"></script>
 
-<script data-sly-test="${wcmmode.isDisabled && project.version}" src="/apps/websight-io/webroot/${project.version}/main.js"></script>
-<script data-sly-test="${wcmmode.isDisabled && !project.version}" src="/apps/websight-io/webroot/main.js"></script>
-<script data-sly-test="${!wcmmode.isDisabled}" src="/apps/websight-io/webroot/main.js"></script>
+<script data-sly-test="${wcmmode.isDisabled && project.version}" src="/apps/kyanite/webroot/${project.version}/main.js"></script>
+<script data-sly-test="${wcmmode.isDisabled && !project.version}" src="/apps/kyanite/webroot/main.js"></script>
+<script data-sly-test="${!wcmmode.isDisabled}" src="/apps/kyanite/webroot/main.js"></script>
 
-<script data-sly-test="${wcmmode.isDisabled && project.version }" src="/apps/websight-io/webroot/${project.version}/main.published.js"></script>
-<script data-sly-test="${wcmmode.isDisabled && !project.version}" src="/apps/websight-io/webroot/main.published.js"></script>
+<script data-sly-test="${wcmmode.isDisabled && project.version }" src="/apps/kyanite/webroot/${project.version}/main.published.js"></script>
+<script data-sly-test="${wcmmode.isDisabled && !project.version}" src="/apps/kyanite/webroot/main.published.js"></script>
